makefile: fix output tests outside of GOPATH
executing the output tests outside of the GOPATH generated the deepequal
artifacts within GOPATH and not the working directory itself.

this commit fixes this by using a temporary directory for generation.
without the temporary directory, the artifacts would be generated into
the directory structure "github.com/cilium/deepequal-gen" within the
working directory.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>